[#170] modify : 커뮤니티 홈 new, hot 컴포넌트 분리
loading spiiner 추가(임시)
