+++ src/pages/community/styles.tsx
@@ -17,7 +17,6 @@ export const CommunityPageContainer = styled.div`
 export const QuestionListWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  /* height: 100%; */
   padding-bottom: 3.125rem;
   .title {
     padding: 0 1.25rem;
